requirements.txt
click>=6.7
mysql-connector>=2.2.9
paramiko>=2.6.0
pyyaml>=3.13